# swagger.json
{
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Documentation of the gotrue API.",
    "title": "gotrue",
    "version": "1.0.0"
  },
  "host": "localhost:9999",
  "basePath": "/",
  "paths": {
...
    "/authorize": {
      "get": {
        "tags": [
          "oauth"
        ],
        "summary": "Redirects the user to the 3rd-party OAuth provider to start the OAuth1.0 or OAuth2.0 authentication process.",
        "operationId": "authorize",
        "parameters": [
          {
            "description": "The redirect url to return the user to after the `/callback` endpoint has completed.",
            "name": "redirect_to",
            "in": "query"
          }
        ],
        "responses": {
          "302": {
            "$ref": "#/responses/authorizeResponse"
          }
        }
      }
    },
...
  },
...
  "responses": {
...
    "authorizeResponse": {
      "description": "Redirects user to the 3rd-party OAuth provider"
    },
...
  },
  "securityDefinitions": {
    "bearer": {
      "type": "apiKey",
      "name": "Authentication",
      "in": "header"
    }
  }
}